💓❣️ Checkpoint ./komutlar/translate.js:26490528/2858 ./atahan.js:26490528/848

+++ komutlar/translate.js
@@ -1,5 +1,5 @@
 const Discord = require("discord.js-selfbot-v13");
-const çeviri = require("google-translate-api")
+const çeviri = require("@iamtraction/google-translate")
 const db = require("quick.db");
 const ayarlar = require("../ayarlar.json");
 let basarili = ayarlar.basariliemoji;
@@ -8,11 +8,33 @@ exports.run = async (client, message, args) => {
   
 if (message.author.id === ayarlar.sahip) {
 
-await message.delete()
-if (!args[0]) return
-if (!args.slice(1).join(' ')) return
+  let mesaj = args.slice(1).join(' ')
+  let dil = args[0]
   
-message.channel.send(çeviri(args.slice(1).join(' '), {to: args[0]}))
+if (!dil) return message.reply(`${basarisiz} ${message.author}, Bir dil belirtin.`).then(x => setTimeout(() => {x.delete()}, 5000))
+
+try {  
+if (dil.toLowerCase() === "aç") {
+  if (await db.fetch("çeviri")) return message.reply(`${basarisiz} ${message.author}, çeviri sistemi zaten açik.`).then(x => setTimeout(() => {x.delete()}, 5000))
+  if (!mesaj) return message.reply(`${basarisiz} ${message.author}, Bir dil belirtin.`).then(x => setTimeout(() => {x.delete()}, 5000))
+  await çeviri("merhaba", {from: "tr", to:mesaj})
+  await db.set("çeviri", `${mesaj}`)
+  message.reply(`${basarili} ${message.author}, çeviri sistemi basariyla açilmistir.`).then(x => setTimeout(() => {x.delete()}, 5000)) 
+  message.react('✅')
+} else if (dil.toLowerCase() === "kapat") {
+  if (!await db.fetch("çeviri")) return message.reply(`${basarisiz} ${message.author}, çeviri sistemi zaten kapali.`).then(x => setTimeout(() => {x.delete()}, 5000))
+  await db.delete("çeviri")
+  message.reply(`${basarili} ${message.author}, çeviri sistemi basariyla kapatilmistir.`).then(x => setTimeout(() => {x.delete()}, 5000)) 
+  message.react('✅')
+} else {
+await message.delete()
+
+if (!mesaj) return message.reply(`${basarisiz} ${message.author}, Mesaj belirtin.`).then(x => setTimeout(() => {x.delete()}, 5000))
+  
+await çeviri(mesaj, {to: dil}).then(x => message.channel.send({content:x.text}))
+message.react('✅')
+}
+} catch { message.reply(`${basarisiz} ${message.author}, Lütfen dogru bir dil girin.`).then(x => setTimeout(() => {x.delete()}, 5000)) }
 }};
 
 exports.conf = {
